# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Order Service API",
    "description": "The API reference of the order service",
    "version": "v0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Orders API",
      "description": "Interface for manipulating orders"
    }
  ],
  "paths": {
    "/api/v1/orders": {
      "get": {
        "tags": [
          "Orders API"
        ],
        "summary": "Finds all orders",
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "Found all",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders API"
        ],
        "summary": "Places an order",
        "operationId": "placeOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Failed to place an order non existent user/duplicate order",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          },
          "201": {
            "description": "Placed an order successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OrderCreationRequest": {
        "required": [
          "email",
          "productId"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "OrderDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          }
        }
      }
    }
  }
}